сортировка по отдельным файлам dashboard
пошла нахуй эта рамка и круглые края, я нахуй ебалась два часа, чтобы нихуя. я ебала, потом сделаю. мне похуй уже,  там остался код нейронки
