+++ grape/src/settingsmenu/settingsmenu.cpp
@@ -1,35 +1,36 @@
-#include "settingsmenu.h"
-#include "ui_settingsmenu.h"
-#include <QMessageBox>
+// #include "settingsmenu.h"
+// #include "ui_settingsmenu.h"
+// #include <QMessageBox>
 
-SettingsMenu::SettingsMenu(QWidget *parent)
-    : QDialog(parent)
-    , ui(new Ui::SettingsMenu)
-{
-    ui->setupUi(this);
-    setWindowFlags(Qt::Popup | Qt::FramelessWindowHint);
-}
+// SettingsMenu::SettingsMenu(QWidget *parent)
+//     : QDialog(parent)
+//     , ui(new Ui::SettingsMenu)
+// {
+//     ui->setupUi(this);
+//     setWindowFlags(Qt::Popup | Qt::FramelessWindowHint);
+// }
 
-SettingsMenu::~SettingsMenu()
-{
-    delete ui;
-}
+// SettingsMenu::~SettingsMenu()
+// {
+//     delete ui;
+// }
 
-void SettingsMenu::on_designButton_clicked()
-{
-    // Будет реализовано позже
-    close();
 
-}
+// void SettingsMenu::on_designButton_clicked()
+// {
+//     // Будет реализовано позже
+//     close();
 
-void SettingsMenu::on_profileButton_clicked()
-{
-    // Будет реализовано позже
-    close();
-}
+// }
 
-void SettingsMenu::on_exitButton_clicked()
-{
-    // Будет реализовано позже
-    close();
-}
+// void SettingsMenu::on_profileButton_clicked()
+// {
+//     // Будет реализовано позже
+//     close();
+// }
+
+// void SettingsMenu::on_exitButton_clicked()
+// {
+//     // Будет реализовано позже
+//     close();
+// }
