+++ grape/src/settingsmenu/settingsmenu.h
@@ -2,6 +2,10 @@
 #define SETTINGSMENU_H
 
 #include <QDialog>
+#include <QPainter>
+#include <QPainterPath>
+#include <QRegion>
+#include <QShowEvent>
 
 namespace Ui {
 class SettingsMenu;
@@ -15,12 +19,20 @@ public:
     explicit SettingsMenu(QWidget *parent = nullptr);
     ~SettingsMenu();
 
+protected:
+    void paintEvent(QPaintEvent *event) override;  // Добавлено
+//     void showEvent(QShowEvent *event) override;    // Добавлено
+//     void resizeEvent(QResizeEvent *event) override;
+
 private slots:
     void on_designButton_clicked();
     void on_profileButton_clicked();
     void on_exitButton_clicked();
 
 private:
+
+    // void updateMask();
+
     Ui::SettingsMenu *ui;
 };
